chore: update versions (#83)
This commit updates a lot of versions. Additionally we *explicitly*
fully specify all dependency versions as we want to update explicitly
rather than have Cargo update implicitly to the latest version. This is
somewhat tied to the resolver behavior, see below.
- `moon`: `1.32.5` --> `1.32.8`
- rust: `1.84.1` --> `1.85.0`
- cargo edition: `2021` --> `2024` see:
[https://doc.rust-lang.org/edition-guide/rust-2024/cargo-resolver.html](https://doc.rust-lang.org/edition-guide/rust-2024/cargo-resolver.html)
- cargo resolver: `2` -> `3` related to the above:
[https://blog.weiznich.de/blog/cargo-instablity/](https://blog.weiznich.de/blog/cargo-instablity/)
- fully specify rust crate dependencies and update to the latest in most
cases. I left the `rasn` dep alone as the new one was published last
night and has fixes for @david_verrus that will allow him to undo some
of his workarounds but I left that as an exercise for him. 😅
- added a dependency on the `toml-cli` binary which allows for `toml`
parsing and incorporated it into the `moon` task that returns the
current rust version (replacing `yq` usage).
- this should close
https://github.com/verrushq/verrus.rs/security/dependabot/1

GitOrigin-RevId: c3394518cb83cb06ae202ece3074fc2090f44089