feat(mms): add full TLS support to command line client  (#84)
We are still not sure if we will be using TLS directly, or making use of
a sidecar, but I wanted to prove out the setup and add support to the
command line client.

1. Added a reasonable default for `Client::TLSConfig` that loads system
certificates.
2. Added full support for TLS with client and server authentication to
the command line client. This should us to interact with the test server
via the Tailscale endpoint. It also provides reference code for other
services seeking to use `rustls` with non-trivial configuration.

GitOrigin-RevId: c167af74fc6248dbcbea700943db9c93b04d4c18