chore(moon): fixes and updates (#92)
Changes and fixes based on the conversation during the tooling
presentation. A few notes:
- we're removing `scripts/thedocker.sh` as it has completely stopped
working properly for the `example-calculator-grpc`. we don't intend this
repo to contain runnable binaries anymore so the `moon` Docker
integration may not be that useful anymore.
- we've moved to explicitly specifying common tasks to be by language or
by language and type. It feels a bit cleaner to be more explicit here
- added comments and updated docs.

GitOrigin-RevId: 8ca236d746ecbed7ddd57bf7808430c924747d64