fix: make it so `rustfmt` will run on stable channel (#114)
With the move to `devenv` we are only able to run a single version of
Rust in the environment and we are running the stable channel (`1.85.0`
currently). This means we cannot invoke the `nightly` toolchain for
`cargo` as it doesn't exist. We do this primarily with `rustfmt` because
a lot of the configuration is still unstable and in the nightly
toolchain. Thus right now we can't properly run formatting locally.

This change updates the `rustfmt.toml` to comment out all the unstable
config so we can rust `rustfmt` locally which means we can run it as a
pre-commit hook (which is also now enabled by this change).

GitOrigin-RevId: 215740a3e62824187d45c0dd1795cd04bb1935ec